JIT optimization: "abandon" invalidated blocks
Invalidated JIT blocks are often invalidated very frequently.
For instance, a SPR opcode is crafted dynamically according to the
sprite data that is to be rendered, which will therefore be overwritten
(and invalidate any JIT cache it belongs to) potentially dozens of times
per frame.
Recompiling a block this way multiple times per frame is very expensive,
and defeats the purpose of a JIT. (That is, execute long straight lines
of host code without overhead.)

To deal with this, by default, do not try to compile+run JIT code for
invalidated blocks, and instead run the basic block through the
interpreter.
An option is provided to override this behavior and force JIT use for
invalidated blocks.

JIT now also outperforms interpreter in ROMs heavy on self-modifying code,
e.g. sf.c16 or raycast.c16.