JIT speedup
Instead of writing to and reading from a bitmask of addresses, write
the invalid bit directly to the basic block struct corresponding to that
address, since it's preallocated.
This simplifies the JIT-emitted invalidation code somewhat, but is more
importantly doing less work in the cpu_rec_validate() function on the
critical path. Previously, profiling revealed it was responsible for
almost 50% of samples with the emulator running at unrestricted speed.
With this patch, cpu_rec_validate is more like 15% of the time. JIT
framerates boosted up to 20-30% on games/demos with very small sprites,
like sf.c16 (Starfield).

Also add profiling option to CMakeLists.